Long ID changed to Integer ID, implements equals, hashcode and toString in Order, Category, Product, User. Tested equals and hashcode in Order

+++ e-shop-plumbing-domain/src/main/java/dao/order/OrderDao.java
@@ -6,7 +6,7 @@ import dao.user.User;
 import java.util.List;
 
 public interface OrderDao {
-    Order selectById(Long id);
+    Order selectById(Integer id);
 
     List<Order> listOrders();
